cart API Integration with profile data fetch

+++ src/main/java/LootlyBackend/repository/ProfileRepo.java
@@ -6,11 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import LootlyBackend.entities.Profile;
+import LootlyBackend.entities.User;
 
 @Repository
 public interface ProfileRepo extends JpaRepository<Profile, Long>{
 	
-	Optional<Profile> findByUserId(Long userId);
+	Optional<Profile> findByUserId(Integer userId);
 
 
 }
